requirements.txt
Flask==3.1.0
Jinja2==3.1.5
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.8.0
